+++ crates/created-swarm/src/swarm.rs
@@ -37,7 +37,7 @@ use futures::stream::iter;
 use hyper::{Body, Request, StatusCode};
 use nox::{Connectivity, Node};
 use particle_protocol::ProtocolConfig;
-use server_config::{default_script_storage_timer_resolution, BootstrapConfig, UnresolvedConfig};
+use server_config::{system_services_config, BootstrapConfig, UnresolvedConfig};
 use test_constants::{EXECUTION_TIMEOUT, KEEP_ALIVE_TIMEOUT, TRANSPORT_TIMEOUT};
 use tokio::sync::oneshot;
 use toy_vms::EasyVM;
@@ -241,10 +241,12 @@ pub struct SwarmConfig {
     pub pool_size: Option<usize>,
     pub builtins_dir: Option<PathBuf>,
     pub spell_base_dir: Option<PathBuf>,
-    pub timer_resolution: Duration,
     pub allowed_binaries: Vec<String>,
-    pub enabled_system_services: Vec<server_config::system_services_config::ServiceKey>,
+    pub enabled_system_services: Vec<String>,
+    pub extend_system_services: Vec<system_services::PackageDistro>,
+    pub override_system_services_config: Option<system_services_config::SystemServicesConfig>,
     pub http_port: u16,
+    pub connector_api_endpoint: Option<String>,
 }
 
 impl SwarmConfig {
@@ -263,10 +265,12 @@ impl SwarmConfig {
             pool_size: <_>::default(),
             builtins_dir: None,
             spell_base_dir: None,
-            timer_resolution: default_script_storage_timer_resolution(),
             allowed_binaries: vec!["/usr/bin/ipfs".to_string(), "/usr/bin/curl".to_string()],
             enabled_system_services: vec![],
+            extend_system_services: vec![],
+            override_system_services_config: None,
             http_port: 0,
+            connector_api_endpoint: None,
         }
     }
 }
@@ -358,9 +362,24 @@ pub fn create_swarm_with_runtime<RT: AquaRuntime>(
     resolved.node_config.aquavm_pool_size = config.pool_size.unwrap_or(1);
     resolved.node_config.particle_execution_timeout = EXECUTION_TIMEOUT;
 
-    resolved.node_config.script_storage_timer_resolution = config.timer_resolution;
     resolved.node_config.allowed_binaries = config.allowed_binaries.clone();
-    resolved.system_services.enable = config.enabled_system_services.clone();
+
+    if let Some(config) = config.override_system_services_config.clone() {
+        resolved.system_services = config;
+    }
+    // `enable_system_services` has higher priority then `enable` field of the SystemServicesConfig
+    resolved.system_services.enable = config
+        .enabled_system_services
+        .iter()
+        .map(|service| {
+            system_services_config::ServiceKey::from_string(service)
+                .expect(&format!("service {service} doesn't exist"))
+        })
+        .collect();
+
+    if let Some(endpoint) = config.connector_api_endpoint.clone() {
+        resolved.system_services.decider.network_api_endpoint = endpoint;
+    }
 
     let management_kp = fluence_keypair::KeyPair::generate_ed25519();
     let management_peer_id = libp2p::identity::Keypair::from(management_kp.clone())
@@ -374,8 +393,21 @@ pub fn create_swarm_with_runtime<RT: AquaRuntime>(
         listen_on: config.listen_on.clone(),
         manager: management_peer_id,
     });
-    let mut node =
-        Node::new(resolved, vm_config, "some version", "some version").expect("create node");
+
+    let system_services_config = resolved.system_services.clone();
+    let system_service_distros =
+        system_services::SystemServiceDistros::default_from(system_services_config)
+            .expect("Failed to get default system service distros")
+            .extend(config.extend_system_services.clone());
+
+    let mut node = Node::new(
+        resolved,
+        vm_config,
+        "some version",
+        "some version",
+        system_service_distros,
+    )
+    .expect("create node");
     node.listen(vec![config.listen_on.clone()]).expect("listen");
 
     (