+++ particle-protocol/src/error.rs
@@ -31,11 +31,12 @@ pub enum ParticleError {
         err: SigningError,
         particle_id: String,
     },
-    #[error("Failed to verify particle {particle_id} signature: {err}")]
+    #[error("Failed to verify particle {particle_id} by {peer_id} with signature: {err}")]
     SignatureVerificationFailed {
         #[source]
         err: VerificationError,
         particle_id: String,
+        peer_id: String,
     },
     #[error("Failed to decode public key from init_peer_id of particle {particle_id}: {err}")]
     DecodingError {